refactor: split git.rs and changelog.rs into submodules
Closes #33
